Changed Settings HustleBenchmarkSeconds and OutlierSeconds to int

+++ Client/Services/SettingsService.cs
@@ -13,8 +13,8 @@ namespace DominosDriverHustleComp.Client.Services
             _serviceProvider = serviceProvider;
         }
 
-        public float HustleBenchmarkSeconds { get; set; }
-        public float OutlierSeconds { get; set; }
+        public int HustleBenchmarkSeconds { get; set; }
+        public int OutlierSeconds { get; set; }
         public int MinDels { get; set; }
         public int MinTrackedPercentage { get; set; }
         public int MaxOverspeeds { get; set; }
