+++ Server/Models/Settings.cs
@@ -6,8 +6,8 @@ namespace DominosDriverHustleComp.Server.Models
     public class Settings
     {
         public int Id { get; set; }
-        public float HustleBenchmarkSeconds { get; set; }
-        public float OutlierSeconds { get; set; }
+        public int HustleBenchmarkSeconds { get; set; }
+        public int OutlierSeconds { get; set; }
         public int MinDels { get; set; }
 
         [Comment("0-100")]
